Bump org.jenkins-ci:process-utils from 1.3 to 1.4 (#68)

+++ src/main/java/org/jenkinsci/test/acceptance/docker/DockerContainer.java
@@ -76,7 +76,7 @@ public class DockerContainer implements Closeable {
         return logfile;
     }
 
-    public int getPid() {
+    public long getPid() {
         return ProcessUtils.getPid(p);
     }
